some fix

+++ routes/api/contacts.js
@@ -10,6 +10,9 @@ const router = express.Router();
 router.get("/", authenticate, async (req, res, next) => {
   try {
     const { page = 1, limit = 10, ...filter } = req.query;
+    if (isNaN(page) || isNaN(limit))
+      throw new CreateError(400, "Page and limit must be a number");
+
     const skip = (page - 1) * limit;
 
     const result = await Contact.find(
@@ -24,7 +27,7 @@ router.get("/", authenticate, async (req, res, next) => {
   }
 });
 
-router.get("/:contactId", async (req, res, next) => {
+router.get("/:contactId", authenticate, async (req, res, next) => {
   try {
     if (!ObjectId.isValid(req.params.contactId))
       throw new CreateError(400, "Not a valid ID");
@@ -32,6 +35,9 @@ router.get("/:contactId", async (req, res, next) => {
     const result = await Contact.findById(req.params.contactId);
 
     if (!result) throw new CreateError(404, "Not found");
+
+    if (String(result.owner) !== req.user.id)
+      throw new CreateError(401, "Not authorized");
 
     res.json(result);
   } catch (e) {
@@ -59,8 +65,14 @@ router.delete("/:contactId", authenticate, async (req, res, next) => {
     if (!ObjectId.isValid(req.params.contactId))
       throw new CreateError(400, "Not a valid ID");
 
+    const result = await Contact.findById(req.params.contactId);
+
+    if (String(result.owner) !== req.user.id)
+      throw new CreateError(401, "Not authorized");
+
     const removedContact = await Contact.findByIdAndDelete(
-      req.params.contactId
+      req.params.contactId,
+      { owner: req.user.id }
     );
 
     if (removedContact) {
@@ -81,6 +93,11 @@ router.put("/:contactId", authenticate, async (req, res, next) => {
     const { error } = schemas.add.validate(req.body);
 
     if (error) throw new CreateError(400, error.message);
+
+    const findedContact = await Contact.findById(req.params.contactId);
+
+    if (String(findedContact.owner) !== req.user.id)
+      throw new CreateError(401, "Not authorized");
 
     const result = await Contact.findByIdAndUpdate(
       req.params.contactId,
@@ -103,6 +120,11 @@ router.patch("/:contactId/favorite", authenticate, async (req, res, next) => {
   const { error } = schemas.updateFavorite.validate(req.body);
 
   if (error) throw new CreateError(400, error.message);
+
+  const findedContact = await Contact.findById(req.params.contactId);
+
+  if (String(findedContact.owner) !== req.user.id)
+    throw new CreateError(401, "Not authorized");
 
   const result = await Contact.findByIdAndUpdate(
     req.params.contactId,
